chore(test): split error-taxonomy contract per package
The per-package unit-test shards introduced in #491 each run only
against ``packages/<package>/tests/`` with that package's own
``pyproject.toml`` rootdir, so the workspace-wide pythonpath that used
to surface ``things_client_fake`` to every shard no longer applies.
``packages/agent-auth/tests/test_error_taxonomy.py`` therefore failed
the agent-auth shard at collection time once Phase 5 of the test.yml
split (#516) re-enabled CI: the file imported ``things_client_fake``
to drive things-bridge's half of the taxonomy, and that module ships
under ``packages/things-bridge/tests/``.

Split the contract along the same per-package boundary that the shards
already enforce. The agent-auth file keeps the agent-auth half and
loses every cross-package import; a sibling file under
``packages/things-bridge/tests/`` carries the things-bridge half and
keeps the ``things_client_fake`` dependency confined to the package
that already ships it. Each docstring cross-references the other so
future readers see the full taxonomy at a glance, and
``scripts/verify-standards.sh`` now grep-checks both files together so
the workspace-wide guarantee (every documented error code is exercised)
stays intact.

Closes #519
